requirements.txt
pillow
requests
pyAesCrypt
pyinstaller
pywin32